PE: Fix expectations of LinkSpecs test on Windows
The output parser detects Windows paths on Unix differently which
makes the format differ from the one that is generated on a
Windows host.
Beside this remove BOM from msvcparser again and do not perform
the format check if there is no expected format specified.
Amends 9e4e9924ad26c708.

Change-Id: Icb0d292a8908cdd2f86b13fde6d2d287f3a96dd6
Reviewed-by: hjk <hjk@qt.io>