Utils: Shortcut Aspect::isDirty for aspects that were never visible
Also do not access Aspect::volatileValue if aspect was never shown

In this case, user-induced unsaved changes cannot exist. This also
avoids QTC_ASSERT in the volatileValue() implementations in these cases.

Silences the contained QTC_ASSERTs properly and avoids returning
a null-variant that's likely not identical to the default value.

Change-Id: Idba89997d0b0b4f9b7dcac0881afe36b35ccdf7c
Reviewed-by: Christian Stenger <christian.stenger@qt.io>