Bump version to 5.0 beta1
Since we almost always do incremental changes instead of major overhauls
in one go, it doesn't make sense to couple changes in the major release
number to major feature overhauls (which is why we've stayed at "4"
since a few years now).

Switch to a version scheme more like semantic versioning. So next
feature release is 5.0, and the following feature-release after 5.0 will
be 6.0, and so on. After all we do break public API in these releases.

Fixes: QTCREATORBUG-25486
Change-Id: Id8510dd75077386e59c537a0c9768f27120a998a
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>