QmlDesigner: Relaxe value property setter
The value.isNull() was misleading. It should only check if a value is
invalid. The null sematics in Qt are special anyway because it they can be
misleading. Most values have no special "null" value like pointers or NaN
with floating point values. QString is for example exposes it's
implementation with a pointer. Sometimes zero is reportet as null but this
not what we wanted to know. So just treat null strings as empty strings.

Change-Id: I6599f6c97f22aff83779a0b32415bf52fd64d82f
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>