Docker: Add a build step
Let's have a build step which allows running a command on a registered
Docker image, which is defined by the Kit's build device.

If the build device is the Local PC, run the command locally.

This specialized build step will most likely be removed after the
existing steps have been taught to run on the Kit's build device.
For now it is part of a POC that allows to experiment with non-local
build commands.

Task-number: QTCREATORBUG-22362
Change-Id: Ifda503ee346b67857a5b82c05cb38cf7f63698d0
Reviewed-by: hjk <hjk@qt.io>