ModelManagerInterface: Cancel all running threads on session switch
On session switch we cancel all the running tasks.
We connect to the SessionManager::aboutToLoadSession()
signal, as this one is emitted just before loading
new session's projects. We don't connect to
SessionManager::aboutToUnloadSession(), since after
this signal is emitted the unloading may be canceled
due to e.g. showing the dialog asking for saving
changed files in the unloaded session (the user may
cancel the process of unloading).

In contrast to what we do on shutdown, we don't wait
for futures being finished here - it's just enough we
cancel all of them.

Fixes: QTCREATORBUG-25583
Change-Id: I01eeca00d150f6e98a80a050c6a19efb848b9954
Reviewed-by: hjk <hjk@qt.io>