Utils: Base TerminalControllingProcess on QtcProcess
This here is a step towards making SynchronousProcess interface more
similar to QtcProcess with the final goal of unifying them.

TerminalControllingProcess is just an implementation details of
SynchronousProcess, which in turn is mostly used in Android and
VcsCommand, e.g. for 'git blame' (good test...)

In the light of 04b20a1a46818b84efbfb11d8cb3c6a8613f608f /
QTCREATORBUG-21547 and the potential need to bisect later this here
is a first baby step.

Change-Id: I48e57d9fdb7c37eb0d2a5c5eef8643d6be97c3cc
Reviewed-by: Orgad Shaneh <orgads@gmail.com>