ProjectExplorer: Whitelist DockerDevice for DesktopRunConfigurations
The idea is that "everything" that can run on the real local host also
can run in a docker container mimicking the same kind of environment.
I.e. all run configurations suitable for current "Desktop" would/should
also be runnable in a docker container, ideally without noticing.

_Desktop_RunConfiguration would then be a bit of a misnomer, it's more
"things that act like 'desktop'", but that can easily be changed later.

Change-Id: I0b284135202e64b54ee86a3815efa95d75a5e29f
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>