Task: Add column information to compiler output tasks
Allows to directly jump to the problematic code
location in the editor.

Done for GCC and Clang which both use the format:
  file:line:column

Not done for MSVC, as this compiler uses the format:
  file(line)
and does not include column information.

It seems like clang-cl does use the format:
  file(line,column)
but the existing parsers were not prepared for
that and were therefore left unchanged for now.

Change-Id: I182634e1c5c941b19801ecafb69fa2b8f91f9d7a
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>