ProjectExplorer: Use run device as fallback for build device
This is not necessarily a permanent solution, but does something
sensible and suppresses warnings about non-existent build devices.

Alternatives would be to always enforce the existence of a build
device (fallback to "Desktop") as done for the run device, or
enforce a build device only and consider "no run device" as
"same as build device.

Change-Id: Ib84670c63a6c416524ab4740419ede6b99981c43
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>