Core: Fix opening files with line column postfix
Having two files with in a locator filter with the same name except one
having an addition line column postfix like
	/tmp/asd
	/tmp/asd+12
made it impossible to open the second file from the locator. Since we
already have the selected file and the locator filter string we also can
use that to parse the postfix manually and make sure to always open the
selected file from the locator widget.

Change-Id: I389603622fbdef01d09b3584192ed6d47e9e170f
Reviewed-by: Christian Stenger <christian.stenger@qt.io>