ProjectExplorer: add option to close all projects except one
This is the project version of the common action present in text
editors. Available when more than 2 projects are open, this option
allows you to close all projects except one that you are interested in.
Quite useful in scenarios where you have been browsing multiple projects
for some research and then want to continue working on one of them.

Change-Id: Ia2b4af4684f5eb7b8f14d5ab9c5af896d492bee2
Reviewed-by: Eike Ziller <eike.ziller@qt.io>