iOS: Don't run event loop while executing xcode-select
This fixes debugging Qt Creator with the iOS plugin enabled. For unclear
reasons running the event loop while executing the process locks Qt
Creator up when debugging. Since running the event loop is a dangerous
hack anyhow and in this case we don't have much benefit from it, just
don't do it.

Change-Id: Ie147d3461823587a987d0920efdfae0839fb5b67
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: hjk <hjk@qt.io>