Locator: Interpret file paths verbatim with "Open and Create"
When typing "f /tmp/test.txt:20" in locator, it would show "Open and
Create /tmp/test.txt:20", but then ask if "/tmp/test.txt" shall be
created.
Make the "Open and Create" take the text verbatim, so it creates "/tmp/
test.txt:20". Specifying a line/column for an empty file doesn't make
sense, and otherwise files with this file name cannot be created.

Change-Id: Ibdd37bfb51518b3077037e7017a759ec0fe116fa
Reviewed-by: David Schulz <david.schulz@qt.io>