Make the eventlisplugin a part of the qmldesigner plugin
The eventlistplugin enables the user to create custom events (QDS-1760).
It provides a dialog to create and remove a custom event or modify its
properties (id, description, shortcut) and a dialog to connect an
event to one of the flowsignals in order to trigger a transition or
a state change.

Task-number: QDS-4300
Change-Id: Ifffa81650d4898b34a65d7901bcb0df95cc397af
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>